neutrix_workspace/prototype/requirements.txt
# --- Core Deep Learning Stack ---
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# --- HuggingFace Donut / Transformers ---
transformers==4.40.0
accelerate>=0.27.0
sentencepiece
protobuf<=4.25.3

# --- OCR and Layout Extraction ---
paddleocr>=2.7.0
paddlepaddle>=2.6.0
opencv-python-headless
PyMuPDF

# --- Document Structure ---
docling
pydantic>=2.0

# --- Utilities ---
pillow
tqdm
pandas
numpy
rapidfuzz
regex
python-dotenv

# --- Visualization / Logging ---
matplotlib
rich

# --- Web Server & Async ---
flask
werkzeug
celery
redis
flasgger
